refactor: rank계산 helper 클래스를 WinningLotto 클래스로 마이그레이션
