+++ src/test/java/lotto/domain/lotto/LottoTicketGroupTest.java
@@ -1,18 +1,13 @@
 package lotto.domain.lotto;
 
-import lotto.domain.rank.Rank;
-import lotto.domain.rank.LottoRankCalculator;
 import org.junit.jupiter.api.Test;
-import lotto.dto.WinLotteryResult;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static lotto.domain.lotto.LottoTicketMockFactory.createLottoHasNumberOneToSix;
 import static lotto.domain.lotto.LottoTicketMockFactory.createLottoTicketGroup;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoTicketGroupTest {
     @Test
@@ -35,29 +30,6 @@ class LottoTicketGroupTest {
 
         // then
         assertThat(result).isEqualTo(lottoTicketList.size());
-    }
-
-
-    @Test
-    void countHitNumber() {
-        // given
-        final LottoTicketGroup lottoTicketGroup = LottoTicketGroup.of(LottoTicketMockFactory.createFourTeenLotto());
-        final LottoTicket winningLotto = createLottoHasNumberOneToSix();
-        final LottoRankCalculator lottoRankCalculator = new LottoRankCalculator();
-        final LottoNumber bonusNumber = LottoNumber.of(LottoNumber.MAX_NO);
-
-        // when
-        final WinLotteryResult result = lottoRankCalculator.calculateWinLotteryResult(lottoTicketGroup, WinningLotto.of(winningLotto, bonusNumber));
-
-        // then
-        assertAll(
-                () -> assertThat(result.getFirstRank().getCount()).isEqualTo(0),
-                () -> assertThat(result.getSecondRank().getCount()).isEqualTo(0),
-                () -> assertThat(result.getThirdRank().getCount()).isEqualTo(0),
-                () -> assertThat(result.getFourthRank().getCount()).isEqualTo(0),
-                () -> assertThat(result.getFifthRank().getCount()).isEqualTo(1),
-                () -> assertThat(result.getTotalPrizeMoney()).isEqualTo(Rank.FIFTH.getWinningMoney())
-        );
     }
 
     @Test
